Allow ssmtp_group to send email
Even with the user in the `sansible_ssmtp_group` group (set as `mail` on my system) I cannot send email apart from as root.
The config files in the /etc/ssmtp directory need read access for the group as well as the owner. Changing them to 640 allows users in the `sansible_ssmtp_group` group to send emails.